Fix the folder structure and styling
