+++ src/app/event/page.jsx
@@ -1,13 +1,15 @@
 "use client";
 
-import CarouselEvent from "@/app/event/CarouselEvent";
-import CardEvent from "@/app/event/CardEvent";
+import CarouselEvent from "@/components/event/CarouselEvent";
+import CardEvent from "@/components/event/CardEvent";
 
 const page = () => {
-  return <div>
-    <CarouselEvent />
-    <CardEvent />
-  </div>;
+  return (
+    <div>
+      <CarouselEvent />
+      <CardEvent />
+    </div>
+  );
 };
 
 export default page;
